Add proper public testing API
